Moved the rest of the custom layers to the layers directory
