+++ src/modified/LevelBrowserLayer.cpp
@@ -3,7 +3,7 @@
 #include <Geode/Bindings.hpp>
 #include "Geode/modify/LevelBrowserLayer.hpp"
 
-#include "GDDLSearchLayer.h"
+#include "layers/GDDLSearchLayer.h"
 
 class $modify(GDDLBrowserLayer, LevelBrowserLayer) {
     struct Fields {
