+++ src/modified/LevelSearchLayer.cpp
@@ -3,7 +3,7 @@
 #include "Geode/modify/LevelSearchLayer.hpp"
 
 #include "RatingsManager.h"
-#include "GDDLSearchLayer.h"
+#include "layers/GDDLSearchLayer.h"
 
 class $modify(GDDLLevelSearchLayer, LevelSearchLayer) {
     // ReSharper disable once CppHidingFunction
